refactor(ui): move view components from routes/ to components/views/
